requirements.txt
akshare>=1.11.0
pandas>=2.0.0
python-telegram-bot>=20.7
requests>=2.31.0
beautifulsoup4>=4.12.0
